chore(billing): refactor createProducts as a dedicated module
