+++ packages/billing/package.json
@@ -46,6 +46,7 @@
     "ajv": "^8.11.0",
     "commander": "^9.4.0",
     "dotenv": "^16.0.1",
+    "p-map": "^5.5.0",
     "ramda": "^0.28.0",
     "stripe": "^10.2.0"
   }
